requirements.txt
keras
tensorflow
pillow
numpy
matplotlib
whatimage
pyheif
Flask
flask-cors
gunicorn
